# apiGateway/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Telemetry API",
    "version": "1.0.0",
    "description": "Read access to GPU telemetry collected by the DCGM-style pipeline (streamer -> message queue -> collector -> ClickHouse)."
  },
  "paths": {
    "/api/v1/gpus": {
      "get": {
        "operationId": "listGPUs",
        "summary": "List GPUs with telemetry",
        "description": "One page of the GPUs that have telemetry, sorted by device id. Follow `next` to walk the rest.",
        "tags": [
          "GPUs"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum rows in the page (1..1000).",
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Zero-based offset to start at. An alternative to cursor for random access; ignored when cursor is present.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "format": "int32"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Resume key taken from a previous page's `next` link. Follow `next` rather than building cursors.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort direction. asc (default) is oldest first; desc is newest first. It also selects which side of the cursor to resume from, so keep it the same while paging.",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GPUListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request: invalid path parameter, window or paging argument. The message names the offending field.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown or expired bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "The query failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable reason; the message names the offending field where one applies"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "GPUListResponse": {
        "properties": {
          "gpus": {
            "items": {
              "properties": {
                "id": {
                  "type": "string",
                  "description": "Stable device identifier (the DCGM GPU UUID); use it in the telemetry path"
                },
                "device": {
                  "type": "string",
                  "description": "NVIDIA device name"
                },
                "index": {
                  "type": "integer",
                  "description": "GPU index on the host. NOT an identifier: use id in the telemetry path"
                },
                "hostname": {
                  "type": "string",
                  "description": "Host the GPU is installed in"
                },
                "model": {
                  "type": "string",
                  "description": "GPU model"
                }
              },
              "additionalProperties": false,
              "type": "object",
              "required": [
                "id",
                "device",
                "index",
                "hostname",
                "model"
              ]
            },
            "type": "array"
          },
          "count": {
            "type": "integer",
            "description": "GPUs in this page (equals limit unless this is the last page)"
          },
          "total": {
            "type": "integer",
            "description": "GPUs that have telemetry"
          },
          "limit": {
            "type": "integer",
            "description": "Page size requested"
          },
          "offset": {
            "type": "integer",
            "description": "Zero-based offset this page started at"
          },
          "order": {
            "type": "string",
            "enum": [
              "asc",
              "desc"
            ],
            "description": "Sort direction by device_id"
          },
          "next": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Absolute URL of the next page; null on the last page"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "gpus",
          "count",
          "total",
          "limit",
          "offset",
          "order",
          "next"
        ],
        "title": "GPUListResponse"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque signed token",
        "description": "Send `Authorization: Bearer <token>`. A long-lived static token works on every route. POST /api/v1/token exchanges one for a short-lived access token; that access token is accepted on the data routes but cannot be exchanged again."
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}